verifactu/requirements.txt
zeep
requests
qrcode
Pillow
unidecode
cryptography==3.4.8